requirements.txt
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
DateTime==5.0
Flask==2.2.2
folium==0.14.0
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
SQLAlchemy==1.4.46
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.11.0
zope.interface==5.5.2
